Target Explorer: Fixed: Delete of static peers

+++ target_explorer/plugins/org.eclipse.tm.te.core.persistence/src/org/eclipse/tm/te/core/persistence/properties/PropertiesFilePersistenceDelegate.java
@@ -70,6 +70,17 @@ public class PropertiesFilePersistenceDelegate extends AbstractPersistenceDelega
 	 */
 	public boolean delete(IPath path) throws IOException {
 		Assert.isNotNull(path);
+
+		// The incoming path has to be an absolute path
+		if (!path.isAbsolute()) {
+			throw new IOException("Not Absolute"); //$NON-NLS-1$
+		}
+
+		// If the file extension is no set, default to "properties"
+		if (path.getFileExtension() == null) {
+			path = path.addFileExtension("properties"); //$NON-NLS-1$
+		}
+
 		return path.toFile().delete();
 	}
